backend/requirements.txt
## If you have trouble downloading/installing PyTorch from PyPI, you can use the official
## PyTorch wheel index by uncommenting the next line:
# --extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.23.0,<0.30.0
python-multipart>=0.0.9,<0.0.10
pydantic>=2.0.0,<3.0.0
opencv-contrib-python>=4.10.0,<5.0.0
numpy>=2.1.0,<3.0.0
Pillow>=10.0.0,<11.0.0
tqdm>=4.66.0,<5.0.0
pytest>=8.0.0,<9.0.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
mediapipe>=0.10.0,<0.11.0
imagehash>=4.3.1,<5.0.0
joblib>=1.0.0,<2.0.0
scikit-learn>=1.5.0,<1.6.0
torch>=1.9.0,<3.0.0
torchvision>=0.10.0,<0.18.0
timm>=0.6.0,<1.0.0
transformers>=4.41.0,<5.0.0
decord>=0.6.0,<0.7.0
pandas>=2.0.0,<3.0.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
boto3>=1.34.0,<2.0.0